refactor(tests): introduce conftest.py
